gpt_disk_io: Remove Drop impl and add io accessor
Once a user has found the partition entry they are looking for, they
may want to read that data. In order to do that, they will need their
BlocKIO device back. This provides a fn (`io`) to get it back.

In order to facilitate the usage of moving `io` out of self, the impl
Drop must be removed. It's up to the user know to implement `flush` on
the drop of their specific Disk.